move files into folders, change path of requires in classes and tests
